backend/requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
resend>=0.8.0
google-generativeai>=0.8.0
reportlab>=4.0.0
drf-spectacular>=0.27.0

# Domain Layer (Local Package)
# En desarrollo local: pip install -e ../domain
# En producción: se instala desde ../domain via build.sh
# litethinking-domain>=1.0.0
